fix(#264): tidy tests
