+++ inc/layout.h
@@ -1,9 +1,6 @@
 #ifndef LAYOUT_H
 #define LAYOUT_H
 
-#include "head.h"
-#include "slist.h"
-
 void layout(void);
 
 #endif // LAYOUT_H
