+++ inc/yaml/marshal.h
@@ -2,7 +2,6 @@
 #define YAML_MARSHAL_H
 
 #include <stdbool.h>
-#include <stddef.h>
 #include <yaml.h>
 
 /*
